+++ aioxmpp/stream.py
@@ -289,14 +289,44 @@ class StanzaToken:
     `on_state_change` may be a function which will be called with the token and
     the new :class:`StanzaState` whenever the state of the token changes.
 
+    .. versionadded:: 0.8
+
+       Stanza tokens are :term:`awaitable`.
+
+    .. describe:: await token
+    .. describe:: yield from token
+
+       Wait until the stanza is either sent or failed to sent.
+
+       :raises ConnectionError: if the stanza enters
+                                :attr:`~.StanzaState.DISCONNECTED` state.
+       :raises RuntimeError: if the stanza enters :attr:`~.StanzaState.ABORTED`
+                             or :attr:`~.StanzaState.DROPPED` state.
+       :return: :data:`None`
+
+       If a coroutine awaiting a token is cancelled, the token is aborted. Use
+       :func:`asyncio.shield` to prevent this.
+
+       .. warning::
+
+          This is no guarantee that the recipient received the stanza. Without
+          stream management, it can not even guarenteed that the server has
+          seen the stanza.
+
+          This is primarily useful as a synchronisation primitive between the
+          sending of a stanza and another stream operation, such as closing the
+          stream.
+
     .. autoattribute:: state
 
     .. automethod:: abort
     """
+    __slots__ = ("stanza", "_state", "on_state_change", "_sent_future")
 
     def __init__(self, stanza, *, on_state_change=None):
         self.stanza = stanza
         self._state = StanzaState.ACTIVE
+        self._sent_future = None
         self.on_state_change = on_state_change
 
     @property
@@ -308,10 +338,29 @@ class StanzaToken:
 
         return self._state
 
+    def _update_future(self):
+        if self._sent_future.done():
+            return
+
+        if self._state == StanzaState.DISCONNECTED:
+            self._sent_future.set_exception(ConnectionError("disconnected"))
+        elif self._state == StanzaState.DROPPED:
+            self._sent_future.set_exception(
+                RuntimeError("stanza dropped by filter")
+            )
+        elif self._state == StanzaState.ABORTED:
+            self._sent_future.set_exception(RuntimeError("stanza aborted"))
+        elif     (self._state == StanzaState.SENT_WITHOUT_SM or
+                  self._state == StanzaState.ACKED):
+            self._sent_future.set_result(None)
+
     def _set_state(self, new_state):
         self._state = new_state
         if self.on_state_change is not None:
             self.on_state_change(self, new_state)
+
+        if self._sent_future is not None:
+            self._update_future()
 
     def abort(self):
         """
@@ -325,10 +374,24 @@ class StanzaToken:
         if     (self._state != StanzaState.ACTIVE and
                 self._state != StanzaState.ABORTED):
             raise RuntimeError("cannot abort stanza (already sent)")
-        self._state = StanzaState.ABORTED
+        self._set_state(StanzaState.ABORTED)
 
     def __repr__(self):
         return "<StanzaToken id=0x{:016x}>".format(id(self))
+
+    @asyncio.coroutine
+    def __await__(self):
+        if self._sent_future is None:
+            self._sent_future = asyncio.Future()
+            self._update_future()
+        try:
+            yield from asyncio.shield(self._sent_future)
+        except asyncio.CancelledError:
+            if self._state == StanzaState.ACTIVE:
+                self.abort()
+            raise
+
+    __iter__ = __await__
 
 
 class StanzaStream:
@@ -2206,49 +2269,15 @@ class StanzaStream:
         """
         Send the given `stanza` over the given :class:`StanzaStream` `stream`.
 
-        Return when the stanza reaches either
-        :attr:`~StanzaState.SENT_WITHOUT_SM` or :attr:`~StanzaState.ACKED`
-        state.
+        .. deprecated:: 0.8
 
-        Raise :class:`ConnectionError` if :attr:`~StanzaState.DISCONNECTED` is
-        reached. Raise :class:`RuntimeError` if :attr:`~StanzaState.ABORTED` or
-        :attr:`~StanzaState.DROPPED` is reached.
-
-        Cancelling the coroutine :meth:`aborts <StanzaToken.abort>` the stanza
-        and returns immediately, even if the stanza has already left the active
-        queue.
+           This functionally equivalent to ``yield from
+           self.enqueue_stanza(stanza)``. Use that instead.
         """
-        fut = asyncio.Future()
-
         self._logger.debug("sending %r and waiting for it to be sent",
                            stanza)
 
-        def cb(token, state):
-            self._logger.debug("token %r enters state %r", token, state)
-            if fut.done():
-                self._logger.warning("state change after future is done!")
-                return
-
-            if state == StanzaState.ACKED:
-                fut.set_result(None)
-            if state == StanzaState.SENT_WITHOUT_SM:
-                fut.set_result(None)
-            if state == StanzaState.DISCONNECTED:
-                fut.set_exception(ConnectionError("disconnected"))
-            if state == StanzaState.DROPPED:
-                fut.set_exception(RuntimeError("stanza dropped by filter"))
-            if state == StanzaState.ABORTED:
-                fut.set_exception(RuntimeError("stanza aborted"))
-
-        token = self.enqueue_stanza(stanza, on_state_change=cb)
-        self._logger.debug("using token %r", token)
-
-        try:
-            yield from fut
-        except asyncio.CancelledError:
-            if token.state == StanzaState.ACTIVE:
-                token.abort()
-            raise
+        yield from self.enqueue_stanza(stanza)
 
 
 @contextlib.contextmanager